+++ dt-connector/src/sinker/kafka/rdkafka_sinker.rs
@@ -77,7 +77,7 @@ impl RdkafkaSinker {
         }
 
         self.base_sinker
-            .update_batch_monitor_for(&task_id, batch_size as u64, data_size as u64)
+            .update_batch_monitor_for(&task_id, batch_size as u64, data_size)
             .await?;
         self.base_sinker.update_monitor_rt_for(&task_id, &rts).await
     }
